added global Formosa theme
In 'Global' theme, you have two time the sequence list of ISO 3166 country codes, followed by a list of profession, that followed by a number from 0 to 63, then a list of 32 places. The result is a culturally neutral sentence associated to a vivid image of a certain professional of a certain country interacting with another professional from another country, for a certain price in a certain place.